refactor: converting server elements to typescript
A production tsconfig is added which extends the original but
overwrites it in order to emit files. Scripts have been added
for running the server in development and production modes.
