+++ Makefile
@@ -2,7 +2,8 @@ SHELL:=/bin/bash
 
 help:
 	@echo "start-dev - Starts webpack-dev-server for local development."
-	@echo "start-production - Starts serving production assets via Express.js."
+	@echo "start-express-dev - Starts serving production assets via Express.js & ts-node-dev (hot reloading upon changes)"
+	@echo "start-production - Starts serving production assets via Express.js & Node.js."
 	@echo "start-lint - Runs linting."
 	@echo "start-types-check - Runs type checking."
 	@echo "start-tests - Runs the test suite."
@@ -10,6 +11,9 @@ help:
 
 start-dev:
 	docker-compose run --rm --service-ports dev
+
+start-express-dev:
+	docker-compose run --rm --service-ports yarn express:dev-server
 
 start-production:
 	docker-compose up web
