+++ package.json
@@ -5,17 +5,21 @@
   "author": "Anthony Hastings <ar.hastings@gmail.com> (https://antwan1986.github.io/)",
   "license": "MIT",
   "private": true,
-  "main": "server/index.mjs",
+  "main": "server/index.ts",
   "scripts": {
     "build": "yarn run webpack:production",
     "browserslist:debug": "browserslist --config=.browserslistrc",
     "browserslist:update": "npx browserslist@latest --update-db && yarn list caniuse-lite",
     "eslint:base": "eslint --ext .js,.mjs,.ts,.tsx --max-warnings 0",
     "eslint:check": "eslint-config-prettier src/components/app/container/index.js",
+    "preexpress:build": "rimraf dist-server/",
+    "express:build": "tsc --project tsconfig.server.json",
+    "express:dev-server": "ts-node-dev server/index.ts",
+    "express:production": "node --experimental-json-modules dist-server/server/index.js",
     "lint": "time npm-run-all --parallel --print-label --continue-on-error lint:*",
     "lint:css": "time stylelint src/**/*.scss",
     "lint:js": "time yarn run eslint:base __mocks__/ server/ src/ support/ webpack/ jest.config.ts",
-    "start": "node --experimental-json-modules server/index.mjs",
+    "start": "yarn run express:production",
     "test": "jest",
     "types:check": "npx tsc --project tsconfig.json",
     "webpack:dev-server": "webpack serve --config=webpack/env.development.ts",
@@ -92,6 +96,7 @@
     "prettier": "2.3.0",
     "react-refresh": "0.10.0",
     "react-test-renderer": "17.0.1",
+    "rimraf": "3.0.2",
     "sass": "1.32.13",
     "sass-loader": "11.1.1",
     "snapshot-diff": "0.8.0",
@@ -100,7 +105,7 @@
     "stylelint-order": "4.1.0",
     "stylelint-scss": "3.19.0",
     "ts-jest": "27.0.2",
-    "ts-node": "10.0.0",
+    "ts-node-dev": "1.1.6",
     "typescript": "4.3.2",
     "webpack": "5.37.0",
     "webpack-bundle-analyzer": "4.4.2",
