+++ tsconfig.server.json
@@ -1,0 +1,12 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "module": "commonjs",
+    "noEmit": false,
+    "outDir": "./dist-server/",
+    "target": "es2015"
+  },
+  "include": [
+    "server/index.ts"
+  ]
+}
